fix: Remove potential memory leaks by exit() include/minishell.h - 106,274: Update header file because of the refactor on minishell.c file and extracting readcmd() to external file.
src/minishell.c
- Extract readcmd() to external directory with new file.
- Add new function looper_wrapper() to wrap lines of code in while loop at main(), to resolve norm error.
- 76: Refactor main() to resolve norm error.

src/minishell_util.c
- Extract getcmd() to external directory with new file.

src/readcmd/readcmd.c
- Move getcmd() and modify the function to return int value depends on the process.
- Add new function looper_readcmd() to fix norm error in previous readcmd() by extracting lines of code in while loop, and return value to control while loop in readcmd().
- Move readcmd() from minishell.c file and modify the function to return error with int -1 to main() if error.

src/executecmd/executecmd.c
- 101: Modify execute_node() to return -1 if execve() went wrong.
